fix: 알림 설정한 공연 개수 쿼리 위치 이동
