+++ app/domain/user-domain/src/main/java/org/example/repository/ticketing/TicketingAlertQuerydslRepository.java
@@ -1,5 +1,9 @@
 package org.example.repository.ticketing;
 
+import java.time.LocalDateTime;
+import java.util.UUID;
+
 public interface TicketingAlertQuerydslRepository {
 
+    long countValidTicketingAlerts(UUID userId, LocalDateTime now);
 }
